Address Namjae review
Includes:

- Add non-atomic add option for linear fv elemental kernels
- Only dispatch if user has implemented a matrix computation routine
- Make Kokkos GeneralUserObject inherit from the coupled dep interface for safety. I think this is
  more consistent because the class already inherits from FESystemHolder
- Remove FaceArgProducer inheritance from Kokkos linear FV. This is a logical change because that
  interface is for the cpu-based functor system, which Kokkos cannot use
- Remove non-ascii
- Remove duplicate kokkos_fe_systems
